treewide: Export symbols needed by Android drivers
We want to enable use of the Android ashmem and binder drivers to
support Anbox, but they should not be built-in as that would waste
resources and increase security attack surface on systems that don't
need them.

Export the currently un-exported symbols they depend on.

(cherry picked from commit eda591e5a88289287362cadefda66f68c580800f)
Signed-off-by: alk3pInjection <webmaster@raspii.tech>