Revert "use lfence instead of rep and nop"
Per a discussion on bbs.archlinux.org [1], LFENCE has a serializing
property on AMD CPUs to block speculative execution. Revert this change
so we can get performance back (if we lost any), on AMD systems.

And how this affects Intel I don't know, but Clearlinux uses this patch
to improve performance.  Perhaps Intel implements LFENCE differently
than AMD?

Also note, the author of this patch didn't include any information for
why this change was made, whether it fixed a bug or improved
performance.

[1] https://bbs.archlinux.org/viewtopic.php?id=235442

This reverts commit 0ba47cad5823f15d37f33df2854f496be55b63fc.

Signed-off-by: alk3pInjection <webmaster@raspii.tech>